refactor(computer-use): extract visual primitives from windows.rs into visual.rs
Split windows.rs (1524L -> 1138L) to satisfy the 1500-line CI file-size
gate. Extracted overlay constants, color helpers, OverlayTone enum, and
low-level Win32 overlay creation/manipulation functions into a new
visual.rs sub-module (440L).

- Moved visual constants (CORNER_*, POINTER_*, CONTROL_*, etc.)
- Moved SESSION_PALETTE + session_color/glow_from_accent/cursor_from_accent
- Moved OverlayTone enum and register_color_overlay_classes
- Moved create_color_overlay/create_cursor_ring_overlay + helpers
- Moved set_overlay_alpha/exclude_overlay_from_capture/position_overlay/etc.
- Added missing Win32 imports to input.rs for test module access

Closes PIP-2792
